FIX: Missing sidebar section link icon for PM tags
No tests cause a missing icon regression is not the end of the world and
I don't feel like it will provide enough value as a regression test long
term.
