+++ app/assets/javascripts/discourse/app/lib/sidebar/user/tags-section/base-tag-section-link.js
@@ -1,0 +1,21 @@
+export default class BaseTagSectionLink {
+  constructor({ tagName }) {
+    this.tagName = tagName;
+  }
+
+  get name() {
+    return this.tagName;
+  }
+
+  get text() {
+    return this.tagName;
+  }
+
+  get prefixType() {
+    return "icon";
+  }
+
+  get prefixValue() {
+    return "tag";
+  }
+}
